song_gem/requirements.txt
google-generativeai>=0.8.0
lyricsgenius>=3.7.0
pandas>=2.0.0
numpy>=1.24.0
textblob>=0.17.0
nltk>=3.8.0
spacy>=3.7.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dateutil>=2.8.0